+++ include/ambition/7kaaInterface/draw.hh
@@ -373,6 +373,11 @@ bool printSaveFilename(
   const int top
 );
 
+bool villagesReportUpperTotals(
+  const int totalVillagers,
+  const int totalPeasants
+);
+
 void printSyncError(
 );